requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gspread==5.1.1
gspread-dataframe==3.2.2
httplib2==0.20.4
idna==3.3
ipykernel==6.8.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.26
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.1.4
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
